backend/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
paho-mqtt==2.1.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
pandas
flask
flask-cors
scikit-learn
